requirements.txt
Flask==3.0.3
gunicorn==20.1.0
google-cloud-aiplatform==1.72.0
google-auth==2.34.0
google-cloud-storage==2.18.2





